requirements_gpu.txt
# GPU版本依赖 - 用于算力平台部署
# 基础数据处理
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
beautifulsoup4>=4.11.0

# 机器学习和深度学习 (GPU版本)
tensorflow-gpu>=2.12.0
# 或者使用统一版本: tensorflow>=2.12.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0

# 股票数据获取
adata>=1.8.0

# Web应用 (可选)
flask>=2.2.0
flask-cors>=4.0.0

# 数据可视化
plotly>=5.10.0

# 性能优化
numba>=0.56.0

# 其他工具
tqdm>=4.64.0
joblib>=1.1.0